backend/requirements/requirements.txt
Django==5.0.6
django-allauth[mfa,socialaccount]==0.63.3
django-otp==1.5.0  # Required for allauth two-factor migration (can be removed later)
django-waffle==4.1.0
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
django-cors-headers==4.3.1
drf-spectacular==0.27.2
stripe==4.2.0
dj-stripe==2.8.4
django-anymail[mailgun]==10.3
django-hijack==3.4.5
django-storages==1.14.3
mailchimp3==3.0.21
requests==2.32.3
sentry-sdk==2.3.1
